+++ examples/rtu-server-address.rs
@@ -14,7 +14,7 @@ struct Service {
 }
 
 impl tokio_modbus::server::Service for Service {
-    type Request = SlaveRequest;
+    type Request = SlaveRequest<'static>;
     type Response = Option<Response>;
     type Error = std::io::Error;
     type Future = future::Ready<Result<Self::Response, Self::Error>>;